Move isV2ClientInputOutputType to is folder
