+++ src/transforms/v2-to-v3/utils/is/index.ts
@@ -1,1 +1,2 @@
 export * from "./isTypeScriptFile";
+export * from "./isV2ClientInputOutputType";
