+++ src/transforms/v2-to-v3/utils/is/isV2ClientInputOutputType.ts
@@ -1,4 +1,4 @@
-import { V2_CLIENT_INPUT_SUFFIX_LIST, V2_CLIENT_OUTPUT_SUFFIX_LIST } from "./config";
+import { V2_CLIENT_INPUT_SUFFIX_LIST, V2_CLIENT_OUTPUT_SUFFIX_LIST } from "../config";
 
 export const isV2ClientInputOutputType = (v2ClientTypeName: string) =>
   [...V2_CLIENT_INPUT_SUFFIX_LIST, ...V2_CLIENT_OUTPUT_SUFFIX_LIST].some((suffix) =>
